pyproject.toml
[project]
name = "gotecq-auto-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.5.3",
    "pytest",
    "python-dotenv>=1.1.1",
    "selenium>=4.34.2",
]
